+++ src/ts/components/volumeslider.ts
@@ -1,5 +1,6 @@
 import {SeekBar, SeekBarConfig} from './seekbar';
 import {UIInstanceManager} from '../uimanager';
+import { VolumeTransition } from '../volumecontroller';
 
 /**
  * Configuration interface for the {@link VolumeSlider} component.
@@ -10,7 +11,7 @@ export interface VolumeSliderConfig extends SeekBarConfig {
    * browser or platform. This currently only applies to iOS.
    * Default: true
    */
-  hideIfVolumeControlProhibited: boolean;
+  hideIfVolumeControlProhibited?: boolean;
 }
 
 /**
@@ -18,9 +19,7 @@ export interface VolumeSliderConfig extends SeekBarConfig {
  */
 export class VolumeSlider extends SeekBar {
 
-  private static readonly issuerName = 'ui';
-
-  constructor(config: SeekBarConfig = {}) {
+  constructor(config: VolumeSliderConfig = {}) {
     super(config);
 
     this.config = this.mergeConfig(config, <VolumeSliderConfig>{
@@ -34,6 +33,8 @@ export class VolumeSlider extends SeekBar {
 
     let config = <VolumeSliderConfig>this.getConfig();
 
+    const volumeController = uimanager.getConfig().volumeController;
+
     if (config.hideIfVolumeControlProhibited && !this.detectVolumeControlAvailability()) {
       this.hide();
 
@@ -42,26 +43,28 @@ export class VolumeSlider extends SeekBar {
       return;
     }
 
-    let volumeChangeHandler = () => {
-      if (player.isMuted()) {
+    volumeController.onChanged.subscribe((_, args) => {
+      if (args.muted) {
         this.setPlaybackPosition(0);
       } else {
-        this.setPlaybackPosition(player.getVolume());
+        this.setPlaybackPosition(args.volume);
       }
-    };
+    });
 
-    player.addEventHandler(player.EVENT.ON_READY, volumeChangeHandler);
-    player.addEventHandler(player.EVENT.ON_VOLUME_CHANGED, volumeChangeHandler);
-    player.addEventHandler(player.EVENT.ON_MUTED, volumeChangeHandler);
-    player.addEventHandler(player.EVENT.ON_UNMUTED, volumeChangeHandler);
+    let volumeTransition: VolumeTransition;
 
+    this.onSeek.subscribe(() => {
+       volumeTransition = volumeController.startTransition();
+    });
     this.onSeekPreview.subscribeRateLimited((sender, args) => {
-      if (args.scrubbing) {
-        player.setVolume(args.position, VolumeSlider.issuerName);
+      if (args.scrubbing && volumeTransition) {
+        volumeTransition.update(args.position);
       }
     }, 50);
     this.onSeeked.subscribe((sender, percentage) => {
-      player.setVolume(percentage, VolumeSlider.issuerName);
+      if (volumeTransition) {
+        volumeTransition.finish(percentage);
+      }
     });
 
     // Update the volume slider marker when the player resized, a source is loaded and player is ready,
@@ -76,8 +79,8 @@ export class VolumeSlider extends SeekBar {
       this.refreshPlaybackPosition();
     });
 
-    // Init volume bar
-    volumeChangeHandler();
+    // Init
+    volumeController.onChangedEvent();
   }
 
   private detectVolumeControlAvailability(): boolean {